feat(table): add expandable example to the cells combined table example
