+++ documentation-site/pages/components/table.mdx
@@ -11,7 +11,6 @@ import Layout from '../../components/layout';
 
 import Basic from 'examples/table/basic.js';
 import Cells from 'examples/table/cells.js';
-import ExpandableCell from 'examples/table/expandable-cell.js';
 import FixedWidthColumn from 'examples/table/fixed-width-column.js';
 import VerticalScroll from 'examples/table/vertical-scroll.js';
 import HorizontalScroll from 'examples/table/horizontal-scroll.js';
@@ -58,13 +57,6 @@ Display larger amounts of information in a scannable UI.
 
 <Example title="Table with column filtering" path="examples/table/filter.js">
   <Filter />
-</Example>
-
-<Example
-  title="Table with expandable cells"
-  path="examples/table/expandable-cell.js"
->
-  <ExpandableCell />
 </Example>
 
 ### Virtualized Table Examples
