+++ src/table/__tests__/table-cells.scenario.js
@@ -16,6 +16,7 @@ import Search from '../../icon/search.js';
 import Plus from '../../icon/plus.js';
 import Delete from '../../icon/delete.js';
 import Overflow from '../../icon/overflow.js';
+import {StatefulPanel} from '../../accordion/index.js';
 
 import {
   StyledTable,
@@ -46,27 +47,220 @@ const StyledLargeText = styled(StyledCell, {
   alignItems: 'center',
 });
 
+const ExpandableCellHead = styled(StyledHeadCell, {
+  paddingTop: '0',
+  paddingBottom: '0',
+  paddingLeft: '0',
+  paddingRight: '0',
+  minWidth: '180px',
+});
+
+const ExpandableCell = styled(StyledCell, {
+  paddingTop: '0',
+  paddingBottom: '0',
+  paddingLeft: '0',
+  paddingRight: '0',
+  minWidth: '180px',
+});
+
+const panelOverrides = {
+  Header: {
+    style: ({$theme: {colors, sizing, typography}, $expanded}) => ({
+      ...typography.font300,
+      color: 'inherit',
+      borderBottom: $expanded
+        ? `1px solid ${colors.mono500}`
+        : '1px solid transparent',
+      paddingTop: sizing.scale300,
+      paddingBottom: sizing.scale300,
+      paddingLeft: sizing.scale1000,
+      paddingRight: sizing.scale600,
+      outline: 'none',
+      ':hover': {
+        color: 'inherit',
+      },
+    }),
+  },
+  Content: {
+    style: ({$theme: {colors, sizing}, $expanded}) => ({
+      backgroundColor: colors.mono200,
+      color: colors.mono800,
+      paddingTop: $expanded ? sizing.scale300 : 0,
+      paddingBottom: $expanded ? sizing.scale600 : 0,
+      paddingLeft: sizing.scale1000,
+      paddingRight: sizing.scale600,
+    }),
+  },
+  ToggleIcon: {
+    style: ({$theme: {colors}}) => ({
+      color: colors.mono600,
+      position: 'absolute',
+      left: '12px',
+    }),
+  },
+};
+
 const DATA = [
-  ['Marlyn', 'Engineering', 'San Francisco', -100, 1234.5],
-  ['Luther', 'Marketing', 'Seattle', 50, 2435.2],
-  ['Kiera', 'Operations', 'Los Angeles', 40, 8348.1],
-  ['Edna', 'Design', 'Atlanta', 700, 2893.4],
-  ['Soraya', 'Finance', 'Denver', 99, 8787.3],
-  ['Dorris', 'Legal', 'Dallas', -20, 6325.2],
-  ['Astrid', 'Product', 'Tempe', 0, 7392.7],
-  ['Wendie', 'Engineering', 'Pittsburgh', -15, 9283.1],
-  ['Marna', 'Marketing', 'Indianapolis', -2, 7720.9],
-  ['Malka', 'Operations', 'New Orleans', 30, 6273.3],
-  ['Jospeh', 'Design', 'New York City', -22, 8837.4],
-  ['Roselee', 'Finance', 'Oakland', 4, 9277.9],
-  ['Justine', 'Legal', 'Louisville', -9, 7737.2],
-  ['Marlon', 'Engineering', 'Baltimore', -2, 2330.3],
-  ['Mellissa', 'Marketing', 'Boulder', -30, 4458.8],
-  ['Fausto', 'Operations', 'Chicago', -10, 6363.9],
-  ['Alfredia', 'Design', 'Grand Rapids', 70, 2235.2],
-  ['Abel', 'Finance', 'Nashville', 30, 9882.3],
-  ['Winford', 'Legal', 'Sacramento', 10, 8774.7],
-  ['Neil', 'Product', 'Columbus', -5, 2673.2],
+  [
+    'Marlyn',
+    'Engineering',
+    'San Francisco',
+    -100,
+    1234.5,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Luther',
+    'Marketing',
+    'Seattle',
+    50,
+    2435.2,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Kiera',
+    'Operations',
+    'Los Angeles',
+    40,
+    8348.1,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Edna',
+    'Design',
+    'Atlanta',
+    700,
+    2893.4,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Soraya',
+    'Finance',
+    'Denver',
+    99,
+    8787.3,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Dorris',
+    'Legal',
+    'Dallas',
+    -20,
+    6325.2,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Astrid',
+    'Product',
+    'Tempe',
+    0,
+    7392.7,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Wendie',
+    'Engineering',
+    'Pittsburgh',
+    -15,
+    9283.1,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Marna',
+    'Marketing',
+    'Indianapolis',
+    -2,
+    7720.9,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Malka',
+    'Operations',
+    'New Orleans',
+    30,
+    6273.3,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Jospeh',
+    'Design',
+    'New York City',
+    -22,
+    8837.4,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Roselee',
+    'Finance',
+    'Oakland',
+    4,
+    9277.9,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Justine',
+    'Legal',
+    'Louisville',
+    -9,
+    7737.2,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Marlon',
+    'Engineering',
+    'Baltimore',
+    -2,
+    2330.3,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Mellissa',
+    'Marketing',
+    'Boulder',
+    -30,
+    4458.8,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Fausto',
+    'Operations',
+    'Chicago',
+    -10,
+    6363.9,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Alfredia',
+    'Design',
+    'Grand Rapids',
+    70,
+    2235.2,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
+  [
+    'Abel',
+    'Finance',
+    'Nashville',
+    30,
+    9882.3,
+    {title: 'Australia', data: '100 Macquarie st. Sydney, Australia'},
+  ],
+  [
+    'Winford',
+    'Legal',
+    'Sacramento',
+    10,
+    8774.7,
+    {title: 'New York', data: '100 Broadway st. New York City, New York'},
+  ],
+  [
+    'Neil',
+    'Product',
+    'Columbus',
+    -5,
+    2673.2,
+    {title: 'California', data: '100 Market st. San Francisco, California'},
+  ],
 ];
 
 export const component = () => (
@@ -77,6 +271,9 @@ export const component = () => (
         <StyledHeadCell>Role</StyledHeadCell>
         <StyledHeadCell>Delta</StyledHeadCell>
         <StyledHeadCell>Amount</StyledHeadCell>
+        <ExpandableCellHead>
+          <StyledHeadCell>Expandable</StyledHeadCell>
+        </ExpandableCellHead>
         <StyledHeadCell>Actions</StyledHeadCell>
       </StyledHead>
       <StyledBody>
@@ -106,6 +303,12 @@ export const component = () => (
                 +1000%
               </Block>
             </StyledLargeText>
+
+            <ExpandableCell>
+              <StatefulPanel title={row[5].title} overrides={panelOverrides}>
+                {row[5].data}
+              </StatefulPanel>
+            </ExpandableCell>
 
             <StyledCell>
               <StyledAction>
